+++ geotrek/common/admin.py
@@ -3,7 +3,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from modeltranslation.admin import TranslationAdmin
 from paperclip.models import Attachment
-from .models import FileType, Organism, Theme, RecordSource
+from . import models as common_models
 
 
 class OrganismAdmin(admin.ModelAdmin):
@@ -54,13 +54,17 @@ class ThemeAdmin(TranslationAdmin):
 
 
 class RecordSourceAdmin(admin.ModelAdmin):
-    list_display = ('name', 'structure', 'pictogram_img')
-    search_fields = ('name', 'structure')
-    list_filter = ('structure',)
+    list_display = ('name', 'pictogram_img')
+    search_fields = ('name', )
 
 
-admin.site.register(Organism, OrganismAdmin)
+class TargetPortalAdmin(admin.ModelAdmin):
+    list_display = ('name', 'website')
+    search_fields = ('name', 'website')
+
+admin.site.register(common_models.Organism, OrganismAdmin)
 admin.site.register(Attachment, AttachmentAdmin)
-admin.site.register(FileType, FileTypeAdmin)
-admin.site.register(Theme, ThemeAdmin)
-admin.site.register(RecordSource, RecordSourceAdmin)
+admin.site.register(common_models.FileType, FileTypeAdmin)
+admin.site.register(common_models.Theme, ThemeAdmin)
+admin.site.register(common_models.RecordSource, RecordSourceAdmin)
+admin.site.register(common_models.TargetPortal, TargetPortalAdmin)