+++ geotrek/core/tests/test_views.py
@@ -164,8 +164,8 @@ class TrailViewsTest(CommonTest):
     @mock.patch('mapentity.models.MapEntityMixin.get_attributes_html')
     def test_document_export(self, get_attributes_html):
         get_attributes_html.return_value = '<p>mock</p>'
-        self.login()
         trail = TrailFactory()
+        self.login()
         with open(trail.get_map_image_path(), 'w') as f:
             f.write('***' * 1000)
         response = self.client.get(trail.get_document_url())