+++ geotrek/flatpages/serializers.py
@@ -2,7 +2,8 @@ from rest_framework import serializers as rest_serializers
 
 from geotrek.flatpages import models as flatpages_models
 from geotrek.common.serializers import (
-    TranslatedModelSerializer, BasePublishableSerializerMixin, RecordSourceSerializer
+    TranslatedModelSerializer, BasePublishableSerializerMixin,
+    RecordSourceSerializer, TargetPortalSerializer
 )
 
 
@@ -11,9 +12,10 @@ class FlatPageSerializer(BasePublishableSerializerMixin, TranslatedModelSerializ
     last_modified = rest_serializers.Field(source='date_update')
     media = rest_serializers.Field(source='parse_media')
     source = RecordSourceSerializer()
+    portal = TargetPortalSerializer()
 
     class Meta:
         model = flatpages_models.FlatPage
         fields = ('id', 'title', 'external_url', 'content', 'target',
-                  'last_modified', 'slug', 'media', 'source') + \
+                  'last_modified', 'slug', 'media', 'source', 'portal') + \
             BasePublishableSerializerMixin.Meta.fields