+++ geotrek/tourism/forms.py
@@ -10,14 +10,14 @@ class TouristicContentForm(CommonForm):
     class Meta:
         fields = ['name', 'category', 'type1', 'type2', 'review', 'published',
                   'description_teaser', 'description', 'themes', 'contact',
-                  'email', 'website', 'practical_info', 'approved', 'source',
+                  'email', 'website', 'practical_info', 'approved', 'source', 'portal',
                   'geom', 'eid', 'reservation_system', 'reservation_id', 'structure', ]
         model = TouristicContent
 
     def __init__(self, *args, **kwargs):
         super(TouristicContentForm, self).__init__(*args, **kwargs)
         # Since we use chosen() in trek_form.html, we don't need the default help text
-        for f in ['themes', 'type1', 'type2', 'source']:
+        for f in ['themes', 'type1', 'type2', 'source', 'portal']:
             self.fields[f].help_text = ''
 
 
@@ -26,11 +26,10 @@ class TouristicEventForm(CommonForm):
 
     class Meta:
         fields = ['name', 'review', 'published', 'description_teaser', 'description',
-                  'themes', 'begin_date', 'end_date', 'duration',
-                  'meeting_point', 'meeting_time', 'contact', 'email',
-                  'website', 'organizer', 'speaker', 'type', 'accessibility',
-                  'participant_number', 'booking', 'target_audience',
-                  'practical_info', 'approved', 'source', 'geom', 'eid', 'structure', ]
+                  'themes', 'begin_date', 'end_date', 'duration', 'meeting_point',
+                  'meeting_time', 'contact', 'email', 'website', 'organizer', 'speaker',
+                  'type', 'accessibility', 'participant_number', 'booking', 'target_audience',
+                  'practical_info', 'approved', 'source', 'portal', 'geom', 'eid', 'structure', ]
         model = TouristicEvent
 
     def __init__(self, *args, **kwargs):